pyproject.toml
[tool.poetry]
name = "dropshot-bot"
version = "0.1.0"
description = ""
authors = ["rvanzelotti <rodrigovanzelotti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
langgraph = "0.0.69"
langchain-openai = "0.1.13"
langchain = "0.2.3"
langchain-community = "0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
